Fixed #1156 - 2.0: Array.count() should return int instead of long (#1166)

+++ shared/src/main/java/com/couchbase/lite/Array.java
@@ -428,7 +428,7 @@ public class Array extends ReadOnlyArray implements ArrayInterface, ObjectChange
     }
 
     private void loadBackingFleeceData() {
-        int count = (int) super.count();
+        int count = super.count();
         for (int i = 0; i < count; i++) {
             Object value = super.getObject(i);
             list.add(CBLData.convert(value, this));
